requirements.txt
beautifulsoup4
urllib3
spotipy
pyyaml
python-dateutil